+++ CHANGELOG.md
@@ -142,6 +142,8 @@ for currentStopLossLimitActivation
 - **Enhancement** Add constructors with parameters to allow custom implementation of ReportGenerators in BacktestExecutor
 - **Enhancement** Added license checker goal on CI's pipeline
 - **Enhancement** Added source format checker goal on CI's pipeline
+- **Enhancement** Added possibility to use CostModels when backtesting with the BacktestExecutor
+- **Enhancement** BacktestExecutor can generate final reports as strings for logs and co.
 
 ### Removed/Deprecated